@@ -105,7 +105,7 @@ class ForumController < ApplicationController
 
   def show
     @forum_post = ForumPost.find(params[:id])
-    @children = ForumPost.paginate :order => "id", :per_page => 30, :conditions => ["parent_id = ?", params[:id]], :page => page_number
+    @children = @forum_post.children.order(:id).paginate :per_page => 30, :page => page_number
 
     if !@current_user.is_anonymous? && @current_user.last_forum_topic_read_at < @forum_post.updated_at
       @current_user.update_attribute(:last_forum_topic_read_at, @forum_post.updated_at)
@@ -115,23 +115,25 @@ class ForumController < ApplicationController
   end
 
   def index
+    @forum_posts = ForumPost.includes(:updater, :creator).order(:is_sticky => :desc, :updated_at => :desc)
     if params[:parent_id]
-      @forum_posts = ForumPost.includes(:updater, :creator).paginate :order => "is_sticky desc, updated_at DESC", :per_page => 100, :conditions => ["parent_id = ?", params[:parent_id]], :page => page_number
+      @forum_posts = @forum_posts.where(:parent_id => params[:parent_id]).paginate :per_page => 100, :page => page_number
     elsif params[:latest]
-      @forum_posts = ForumPost.includes(:updater, :creator).where(:parent_id => nil).order("updated_at DESC").paginate(:page => 1, :per_page => 10)
+      @forum_posts = @forum_posts.where(:parent_id => nil).order(:updated_at => :desc).paginate(:page => 1, :per_page => 10)
     else
-      @forum_posts = ForumPost.includes(:updater, :creator).paginate :order => "is_sticky desc, updated_at DESC", :per_page => 30, :conditions => "parent_id IS NULL", :page => page_number
+      @forum_posts = @forum_posts.where(:parent_id => nil).paginate :per_page => 30, :page => page_number
     end
 
     respond_to_list("forum_posts")
   end
 
   def search
+    @forum_posts = ForumPost.includes(:creator, :updater, :parent).order(:id => :desc)
     if params[:query]
       query = params[:query].scan(/\S+/).join(" & ")
-      @forum_posts = ForumPost.includes(:creator, :updater, :parent).paginate :order => "forum_posts.id DESC", :per_page => 30, :conditions => ["forum_posts.text_search_index @@ plainto_tsquery(?)", query], :page => page_number
+      @forum_posts = @forum_posts.where("text_search_index @@ plainto_tsquery(?)", query).paginate :per_page => 30, :page => page_number
     else
-      @forum_posts = ForumPost.includes(:creator, :updater, :parent).paginate :order => "forum_posts.id DESC", :per_page => 30, :page => page_number
+      @forum_posts = @forum_posts.paginate :per_page => 30, :page => page_number
     end
 
     respond_to_list("forum_posts")
