@@ -338,7 +338,7 @@ class PoolController < ApplicationController
     def zip
       # FIXME: should use the correct mime type instead of this hackery.
       Rack::MiniProfiler.deauthorize_request if Rails.env.development?
-      pool = Pool.find(params[:id], :include => [:pool_posts => :post])
+      pool = Pool.includes(:pool_posts => :post).find(params[:id])
       @pool_zip = pool.get_zip_data(params)
       headers["X-Archive-Files"] = "zip"
       render :layout => false
