@@ -33,7 +33,7 @@ class PostController < ApplicationController
   end
 
   def create
-    if @current_user.is_member_or_lower? && Post.count(:conditions => ["user_id = ? AND created_at > ? ", @current_user.id, 1.day.ago]) >= CONFIG["member_post_limit"]
+    if @current_user.is_member_or_lower? && Post.where(:user_id => @current_user.id).where("created_at > ?", 1.day.ago).count >= CONFIG["member_post_limit"]
       respond_to_error("Daily limit exceeded", { :action => "error" }, :status => 421)
       return
     end
