@@ -77,8 +77,8 @@ class Pool < ActiveRecord::Base
     end
 
     def transfer_post_to_parent(post_id, parent_id)
-      pool_post = pool_posts.find(:first, :conditions => ["post_id = ?", post_id])
-      parent_pool_post = pool_posts.find(:first, :conditions => ["post_id = ?", parent_id])
+      pool_post = pool_posts.find_by(:post_id => post_id)
+      parent_pool_post = pool_posts.find_by(:post_id => parent_id)
       return unless parent_pool_post.nil?
 
       sequence = pool_post.sequence
