@@ -1,8 +1,10 @@
 module Post::SqlMethods
   module ClassMethods
     def find_by_tag_join(tag, options = {})
-      tag = tag.downcase.tr(" ", "_")
-      find(:all, :conditions => ["tags.name = ?", tag], :select => "posts.*", :joins => "JOIN posts_tags ON posts_tags.post_id = posts.id JOIN tags ON tags.id = posts_tags.tag_id", :limit => options[:limit], :offset => options[:offset], :order => (options[:order] || "posts.id DESC"))
+      joins(:_tags).where(:tags => { :name => tag.downcase.tr(" ", "_") })
+        .limit(options[:limit])
+        .offset(options[:offset])
+        .order(options[:order] || { :id => :desc })
     end
 
     def sql_range_for_where(parsed_query, field)
@@ -44,8 +46,6 @@ module Post::SqlMethods
         c << "#{field} IN (#{items.join(", ")})"
         p.concat(arr[1])
 
-      else
-        # do nothing
       end
     end
 
