@@ -15,7 +15,7 @@ Moebooru::Application.configure do
   config.action_controller.perform_caching = true
 
   # Disable Rails's static asset server (Apache or nginx will already do this)
-  config.serve_static_assets = false
+  config.serve_static_files = false
 
   # Compress JavaScripts and CSS
   config.assets.js_compressor = :uglifier
@@ -82,5 +82,4 @@ Moebooru::Application.configure do
                           :sender_address => ("%s <%s>" % ["notifier", CONFIG["email_from"]]),
                           :exception_recipients => CONFIG["admin_contact"]
                         }
-
 end
